[FLINK-30751] [docs] Remove references to disableDataSync in RocksDB documentation
Remove confusing references to syncing to disk, which appear to be leftover from mentioning that setUseFsync is false. But this is true for all predefined options, as mentioned at the top of the doc. We should remove this for clarity, and can mention it elsewhere a single time.

Lexicographically sort the order of options in both documentation and code for improved readability and maintainability.

Reword the introduction to mention the setUseFsync option being false, irrespective of the PredefinedOptions setting. Add the explanation of why syncing is not necessary in Flink to the introduction, where it has better context, and remove the leftover reference to syncing from the DEFAULT.