[FLINK-33346][runtime][test] Removes timeout
The instability is caused by the timeout. The fatal error hides the actual cause (i.e. the TimeoutException). It can be reproduced by setting the timeout of the test to 0ms. The actual cause is revealed when disabling the fatal error handling or not throwing a AssertionError in case of the fatal error.